refactor: redux
