+++ src/http-common.ts
@@ -6,3 +6,4 @@ export default axios.create({
     "Content-type": "application/json"
   }
 });
+
