+++ src/redux/usersSlice.ts
@@ -1,85 +1,13 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import axios from 'axios';
+import { createSlice } from '@reduxjs/toolkit';
+import { initialState } from '../types/type';
+import {
+  fetchUsers,
+  createUser,
+  fetchUserById,
+  updateUser,
+  deleteUser,
+} from './thunks';
 
-
-// Định nghĩa kiểu User
-interface User {
-  id: number;
-  email: string;
-  first_name: string;
-  last_name: string;
-  avatar: string;
-}
-
-// Định nghĩa kiểu cho state
-interface UsersState {
-  users: User[]; // Danh sách người dùng
-  currentUser: User | null; // Người dùng hiện tại
-  loading: boolean;
-  error: string | null;
-
-  totalPages: number; // Thêm trường totalPages
-}
-
-// Khởi tạo state ban đầu
-const initialState: UsersState = {
-  users: [], // Khởi tạo với mảng rỗng
-  currentUser: null, // Không có người dùng hiện tại
-  loading: false,
-  error: null,
-  totalPages: 0, // Khởi tạo với 0
-};
-
-// Tạo async thunk để lấy danh sách người dùng theo trang
-// export const fetchUsers = createAsyncThunk('users/fetchUsers', async (page: number) => {
-//   const response = await axios.get(`https://reqres.in/api/users?page=${page}`);
-//   return response.data.data;
-// });
-
-
-export const fetchUsers = createAsyncThunk(
-  'users/fetchUsers',
-  async (page: number) => {
-    const response = await axios.get(`https://reqres.in/api/users?page=${page}`);
-    return response.data; // Trả về toàn bộ dữ liệu từ API
-  }
-);
-
-
-// ----------
-
-// Tạo async thunk để lấy danh sách người dùng
-// export const fetchUsers = createAsyncThunk('users/fetchUsers', async () => {
-//   const response = await axios.get('https://reqres.in/api/users');
-//   return response.data.data;
-// });
-
-// Tạo async thunk để tạo người dùng
-export const createUser = createAsyncThunk('users/createUser', async (user: User) => {
-  const response = await axios.post('https://reqres.in/api/users', user);
-  return response.data;
-});
-
-// Tạo async thunk để lấy người dùng theo ID
-export const fetchUserById = createAsyncThunk('users/fetchUserById', async (id: number) => {
-  const response = await axios.get(`https://reqres.in/api/users/${id}`);
-  console.log("detail thanh cong response.data.data", response.data.data);
-
-  return response.data.data;
-});
-
-
-// Tạo async thunk để cập nhật người dùng
-export const updateUser = createAsyncThunk('users/updateUser', async (user: User) => {
-  const response = await axios.put(`https://reqres.in/api/users/${user.id}`, user);
-  return response.data;
-});
-
-// Tạo async thunk để xóa người dùng
-export const deleteUser = createAsyncThunk('users/deleteUser', async (id: number) => {
-  await axios.delete(`https://reqres.in/api/users/${id}`);
-  return id; // Trả về ID để xóa
-});
 
 const usersSlice = createSlice({
   name: 'users',
@@ -94,13 +22,13 @@ const usersSlice = createSlice({
   },
   extraReducers: (builder) => {
     builder
-      // list
+      // Fetch Users
       .addCase(fetchUsers.pending, (state) => {
         state.loading = true; // Khi bắt đầu fetch, set loading thành true
-        state.error = null; // Reset lỗi
+        state.error = null;  // Reset lỗi
       })
       .addCase(fetchUsers.fulfilled, (state, action) => {
-        state.loading = false; // Khi fetch thành công, set loading thành false
+        state.loading = false;  // Khi fetch thành công, set loading thành false
         state.users = action.payload.data; // Lưu danh sách người dùng
         state.totalPages = action.payload.total_pages; // Cập nhật totalPages
       })
@@ -108,46 +36,28 @@ const usersSlice = createSlice({
         state.loading = false; // Khi fetch thất bại, set loading thành false
         state.error = action.error.message || 'Failed to fetch users'; // Cập nhật lỗi
       })
-
-      // create
+      // Create User
       .addCase(createUser.fulfilled, (state, action) => {
         state.users.push(action.payload);
       })
-
-      // UserById
-      // .addCase(fetchUserById.pending, (state) => {
-      //   state.loading = true;
-      // })
-      // .addCase(fetchUserById.fulfilled, (state, action) => {
-      //   state.loading = false;
-      //   state.users = action.payload;
-      // })
-      // .addCase(fetchUserById.rejected, (state, action) => {
-      //   state.loading = false;
-      //   state.error = action.error.message || 'Failed to fetch user';
-      // })
-
-      // Xử lý các action khác như fetchUsers, fetchUserById...
-      // .addCase(fetchUsers.fulfilled, (state, action) => {
-      //   state.loading = false;
-      //   state.users = action.payload; // Cập nhật danh sách người dùng
-      // })
+      // Fetch User By ID
       .addCase(fetchUserById.fulfilled, (state, action) => {
         state.loading = false;
-        state.currentUser = action.payload; // Cập nhật người dùng hiện tại
+        state.currentUser = action.payload;
       })
-
-      // update
+      // Update User
       .addCase(updateUser.fulfilled, (state, action) => {
-        state.users = action.payload;
+        const index = state.users.findIndex(user => user.id === action.payload.id);
+        if (index !== -1) {
+          state.users[index] = action.payload; // Cập nhật người dùng
+        }
       })
-
-      // delete
-
+      // Delete User
       .addCase(deleteUser.fulfilled, (state, action) => {
         state.users = state.users.filter(user => user.id !== action.payload);
       });
   },
 });
+
 export const { setCurrentUser, clearCurrentUser } = usersSlice.actions;
 export default usersSlice.reducer;
