+++ src/services/UserService.ts
@@ -1,21 +1,23 @@
 import http from "../http-common";
 
-import IUserlData from "../types/User";
+import User from "../types/User";
 
-const getAll = () => {
-  return http.get<Array<IUserlData>>("/users");
+// Hàm mới để lấy danh sách người dùng theo trang
+const getAllByPage = (page: number) => {
+  return http.get<Array<User>>(`/users?page=${page}`);
 };
+
 
 // detail
 const get = (id: any) => {
-  return http.get<IUserlData>(`/users/${id}`);
+  return http.get<User>(`/users/${id}`);
 };
 
-const create = (data: IUserlData) => {
-  return http.post<IUserlData>("/users", data);
+const create = (data: User) => {
+  return http.post<User>("/users", data);
 };
 
-const update = (id: any, data: IUserlData) => {
+const update = (id: any, data: User) => {
   return http.put<any>(`/users/${id}`, data);
 };
 
@@ -23,18 +25,21 @@ const remove = (id: any) => {
   return http.delete<any>(`/users/${id}`);
 };
 
+
+const getAll = () => {
+  return http.get<Array<User>>("/users");
+};
+
+
 const removeAll = () => {
   return http.delete<any>(`/users`);
 };
 
 const findByTitle = (title: string) => {
-  return http.get<Array<IUserlData>>(`/users?title=${title}`);
+  return http.get<Array<User>>(`/users?title=${title}`);
 };
 
-// Hàm mới để lấy danh sách người dùng theo trang
-const getAllByPage = (page: number) => {
-  return http.get<Array<IUserlData>>(`/users?page=${page}`);
-};
+
 
 const userservice = {
   getAll,
